+++ app/core/functions.php
@@ -1,0 +1,7 @@
+<?php
+
+function show($stuff){
+    echo "<pre>";
+    print_r($stuff);
+    echo "</pre>";
+}
